Rename IconTooltip to HelpTooltip as we dont need the icon support for now
