+++ packages/react-app/src/components/field-input/help-tooltip.tsx
@@ -1,0 +1,60 @@
+import { Help, useTheme } from '@1hive/1hive-ui';
+import { ReactNode, useRef } from 'react';
+import { GUpx } from 'src/utils/style.util';
+import styled from 'styled-components';
+
+// #region Styled
+
+const TooltipWrapperStyled = styled.div`
+  color: #637381;
+`;
+
+const HelpWrapperStyled = styled.div`
+  margin-left: ${GUpx()};
+  margin-right: ${GUpx(0.5)};
+  svg {
+    color: ${({ theme }: any) => theme.hint}!important;
+  }
+  button:focus::after {
+    border: none;
+  }
+`;
+
+// #endregion
+
+type Props = {
+  tooltip?: ReactNode;
+  children?: ReactNode;
+};
+
+export const HelpTooltip = ({ tooltip, children }: Props) => {
+  const theme = useTheme();
+  const wrapperRef = useRef<HTMLElement>(null);
+
+  const handleEnter = () => {
+    // Simulate help button click
+    const btn = wrapperRef.current?.getElementsByTagName('button')[0];
+    btn?.click();
+  };
+  const handleLeave = () => {
+    // Simulate ESC key press
+    document.dispatchEvent(new KeyboardEvent('keydown', { keyCode: 27 }));
+  };
+
+  return (
+    <>
+      <HelpWrapperStyled
+        theme={theme}
+        onMouseEnter={handleEnter}
+        onMouseLeave={handleLeave}
+        ref={wrapperRef}
+      >
+        <Help hint={tooltip}>
+          <TooltipWrapperStyled color={theme.accentContent}>
+            {children ?? tooltip}
+          </TooltipWrapperStyled>
+        </Help>
+      </HelpWrapperStyled>
+    </>
+  );
+};
